Crud, Rota, e EndPoitns de Coleção parcial

crud, rota feitas
endpoint de:
addVia a uma colecao
removerVia de uma colecao
encontrar colecoes do usuario
encontrar colecao do usuario
Ajustes também feitos em:

mudança de colecaoEscalada para Escalada
mudança em Usuario que agora só contem os próprios atributos relacionados a ele
mudança nos scripts do bd para estar de acordo com ajustes

+++ backend/Application/services/UsuarioService.ts
@@ -1,5 +1,5 @@
-import {UsuarioRepository} from "../../Infrastructure/repositories/UsuarioRepository";
 import {Usuario} from "../../Domain/models/Usuario";
+import {UsuarioRepository} from "../../Infrastructure/repositories/UsuarioRepository";
 
 
 export class UsuarioService {
@@ -22,10 +22,16 @@ export class UsuarioService {
     }
 
     async updateUsuario(usuario: Usuario): Promise<void> {
+        if (!await this.getUsuarioById(usuario.id)) {
+            throw new Error("Usuario não encontrada");
+        }
         return this.repository.updateUsuario(usuario);
     }
 
     async deleteUsuario(id: number): Promise<void> {
+        if (!await this.getUsuarioById(id)) {
+            throw new Error("Usuario não encontrada");
+        }
         return this.repository.deleteUsuario(id);
     }
 }
